requirements.txt
jupyter-book>=0.15.1
PyYAML>=6.0
GitPython>=3.1.31
sphinx-book-theme>=1.0.1
sphinx-copybutton>=0.5.2
sphinx-togglebutton>=0.3.2
myst-nb>=0.17.2
sphinx-design>=0.5.0